requirements.txt
asgiref==3.5.2
Django==4.1.2
django-autoslug==1.9.8
Pillow==9.2.0
sqlparse==0.4.3
tzdata==2022.4
